+++ Twig/PaymentLink.php
@@ -9,21 +9,21 @@
 
 namespace c975L\PaymentBundle\Twig;
 
-use Twig_Environment;
-use Twig_Extension;
-use Twig_SimpleFunction;
+use Twig\Environment;
+use Twig\Extension\AbstractExtension;
+use Twig\TwigFunction;
 
 /**
  * Twig extension to display the Payment button using `payment_link(['YOUR_TEXT_TO_DISPLAY', AMOUNT, 'CURRENCY'])`
  * @author Laurent Marquet <laurent.marquet@laposte.net>
  * @copyright 2018 975L <contact@975l.com>
  */
-class PaymentLink extends Twig_Extension
+class PaymentLink extends AbstractExtension
 {
     public function getFunctions()
     {
         return array(
-            new Twig_SimpleFunction(
+            new TwigFunction(
                 'payment_link',
                 array($this, 'paymentLink'),
                 array(
@@ -38,7 +38,7 @@ class PaymentLink extends Twig_Extension
      * Returns xhtml code for Payment link
      * @return string
      */
-    public function paymentLink(Twig_Environment $environment, $text = null, $amount = null, $currency = null)
+    public function paymentLink(Environment $environment, $text = null, $amount = null, $currency = null)
     {
         return $environment->render('@c975LPayment/fragments/paymentLink.html.twig', array(
                 'text' => $text,